Step 3: Modularize into ascii_art.py + game_logic.py; enhance loop with validated guesses, win/lose conditions
