+++ ascii.py
@@ -1,0 +1,24 @@
+STAGES = [
+    """
+      ___  
+     /___\\ 
+     (o o) 
+     ( : ) 
+     ( : ) 
+    """,
+    """
+         ___  
+        /___\\ 
+        (o o) 
+        ( : ) 
+       """,
+    """
+         ___  
+        /___\\ 
+        (o o) 
+       """,
+    """
+      ___  
+     /___\\ 
+    """
+]
